making file object in line with readme
The readme states that the file object returned by the server can use any or all of the fields, or even define its own. This fixes a bug that assumed the existence of all file object fields as well as their order, and allows arbitrary additional fields to be specified by the server.